demo

+++ src/quanter_swarm/evaluation/sharpe.py
@@ -10,4 +10,4 @@ def sharpe_ratio(returns: list[float]) -> float:
     variance = sum((value - avg) ** 2 for value in returns) / len(returns)
     if variance == 0:
         return round(avg, 4)
-    return round(avg / (variance ** 0.5) * sqrt(len(returns)), 4)
+    return float(round(avg / (variance ** 0.5) * sqrt(len(returns)), 4))
